requirements.txt
fastapi
uvicorn[standard]
pillow
transformers
torch
pymongo
python-multipart
requests
